@@ -1,72 +1,69 @@
 {
-    "author": "Charbel Chahine",
-    "email": "chahine.charbel@me.com",
-    "license": "MIT",
-    "name": "mountaincentral",
-    "description": "Multilingual Gatsby Starter",
-    "version": "2.1.0",
+    "name": "mountain-central",
+    "private": true,
+    "description": "A simple starter to get up and developing quickly with Gatsby. It supports multiple languages, NetlifyCMS, styling with Sass, Material-UI components & dark mode",
+    "version": "3.0.0",
+    "author": "Charbel Chahine <chahine.charbel@me.com>",
     "dependencies": {
-        "@babel/core": "^7.9.6",
-        "@material-ui/core": "^4.4.1",
-        "@material-ui/icons": "^4.4.1",
-        "babel-jest": "^24.9.0",
-        "dompurify": "^2.0.11",
-        "enzyme": "^3.10.0",
-        "enzyme-adapter-react-16": "^1.14.0",
-        "eslint": "^6.2.2",
-        "eslint-config-airbnb": "^18.0.1",
-        "eslint-config-prettier": "^6.1.0",
-        "eslint-plugin-import": "^2.18.2",
-        "eslint-plugin-jest": "^22.16.0",
-        "eslint-plugin-jsx-a11y": "^6.2.3",
-        "eslint-plugin-prettier": "^3.1.0",
-        "eslint-plugin-react": "^7.14.3",
-        "gatsby": "^2.21.19",
-        "gatsby-cli": "^2.12.10",
-        "gatsby-core-utils": "^1.2.1",
-        "gatsby-image": "^2.4.3",
-        "gatsby-plugin-catch-links": "^2.3.1",
+        "@material-ui/core": "^4.9.14",
+        "@material-ui/icons": "^4.9.1",
+        "babel-jest": "^26.0.1",
+        "gatsby": "^2.21.22",
+        "gatsby-image": "^2.4.5",
         "gatsby-plugin-manifest": "^2.4.2",
-        "gatsby-plugin-netlify-cms": "^4.3.1",
-        "gatsby-plugin-netlify-cms-paths": "^1.3.0",
-        "gatsby-plugin-nprogress": "^2.3.1",
+        "gatsby-plugin-netlify-cms": "^4.3.3",
         "gatsby-plugin-offline": "^3.2.1",
         "gatsby-plugin-react-helmet": "^3.3.1",
         "gatsby-plugin-react-svg": "^3.0.0",
-        "gatsby-plugin-sass": "^2.3.1",
-        "gatsby-plugin-sharp": "^2.6.2",
-        "gatsby-plugin-sitemap": "^2.4.2",
+        "gatsby-plugin-sass": "^2.3.2",
+        "gatsby-plugin-sharp": "^2.6.9",
         "gatsby-plugin-use-dark-mode": "^1.1.2",
-        "gatsby-remark-autolink-headers": "^2.3.1",
-        "gatsby-remark-images": "^3.3.1",
+        "gatsby-remark-images": "^3.3.8",
         "gatsby-source-filesystem": "^2.3.1",
-        "gatsby-source-stripe": "^3.0.7",
-        "gatsby-transformer-remark": "^2.8.7",
-        "gatsby-transformer-sharp": "^2.5.2",
-        "graphql": "^15.0.0",
-        "html-react-parser": "^0.10.3",
-        "i18n-react": "^0.7.0",
-        "jest": "^26.0.1",
-        "netlify-cms-app": "^2.12.12",
-        "node-sass": "^4.13.0",
-        "prettier": "^1.18.2",
+        "gatsby-transformer-remark": "^2.8.13",
+        "gatsby-transformer-sharp": "^2.5.3",
+        "i18next": "^19.4.4",
+        "netlify-cms-app": "^2.12.13",
+        "node-sass": "^4.14.1",
         "prop-types": "^15.7.2",
-        "react": "^16.9.0",
-        "react-dom": "^16.9.0",
-        "react-helmet": "^5.2.1",
-        "react-hot-loader": "^4.12.21",
-        "react-test-renderer": "^16.9.0",
-        "slugify": "^1.3.4",
-        "use-dark-mode": "^2.3.1",
-        "webpack": "^4.43.0"
+        "react": "^16.12.0",
+        "react-dom": "^16.12.0",
+        "react-helmet": "^6.0.0",
+        "react-i18next": "^11.4.0",
+        "react-markdown": "^4.3.1",
+        "use-dark-mode": "^2.3.1"
     },
+    "devDependencies": {
+        "enzyme": "^3.11.0",
+        "enzyme-adapter-react-16": "^1.15.2",
+        "eslint": "^7.0.0",
+        "eslint-config-airbnb": "^18.1.0",
+        "eslint-config-prettier": "^6.11.0",
+        "eslint-plugin-import": "^2.20.2",
+        "eslint-plugin-jest": "^23.13.1",
+        "eslint-plugin-jsx-a11y": "^6.2.3",
+        "eslint-plugin-prettier": "^3.1.3",
+        "eslint-plugin-react": "^7.20.0",
+        "jest": "^26.0.1",
+        "prettier": "2.0.5"
+    },
+    "license": "MIT",
     "scripts": {
-        "start": "gatsby develop",
         "build": "gatsby build",
+        "develop": "gatsby develop",
+        "format": "prettier --write \"**/*.{js,jsx,json,md}\"",
+        "start": "gatsby develop",
         "serve": "gatsby serve",
-        "test": "jest --coverage",
+        "clean": "gatsby clean",
+        "test": "jest",
         "jest:w": "jest --watch",
         "sass": "sass --watch src/styles/scss/index.scss:src/styles/css/index.css"
     },
-    "homepage": "https://mountaincentral.netlify.com"
+    "repository": {
+        "type": "git",
+        "url": "https://github.com/gatsbyjs/gatsby-starter-default"
+    },
+    "bugs": {
+        "url": "https://github.com/gatsbyjs/gatsby/issues"
+    }
 }
