@@ -1,4 +1,8 @@
 /**
- * Gatsby's Browser APIs.
- * https://www.gatsbyjs.org/docs/browser-apis/
- * */
+ * Implement Gatsby's Browser APIs in this file.
+ *
+ * See: https://www.gatsbyjs.org/docs/browser-apis/
+ */
+
+// eslint-disable-next-line import/prefer-default-export
+export { wrapPageElement } from './gatsby-ssr'
