@@ -1,95 +1,73 @@
 /**
- * Gatsby's Node APIs.
- * https://www.gatsbyjs.org/docs/node-apis/
- * */
+ * Implement Gatsby's Node APIs in this file.
+ *
+ * See: https://www.gatsbyjs.org/docs/node-apis/
+ */
 
-const path = require('path');
-const slugify = require('slugify');
-const i18n = require('./src/i18n/config/i18n');
+/**
+ * Creates localized paths for each file in the /pages folder.
+ * For example, pages/404.js will be converted to /en/404.js and /el/404.js and
+ * it will be accessible from https:// .../en/404/ and https:// .../el/404/
+ */
+const { currentLanguages } = require('./src/i18n/config/currentLanguages')
 
-const Promise = require(`bluebird`);
+exports.onCreatePage = async ({ page, actions: { createPage, deletePage } }) => {
+    // Delete the original page (since we are gonna create localized versions of it)
+    await deletePage(page)
 
-exports.onCreatePage = ({ page, actions }) => {
-    if (page.componentPath && page.componentPath.match(/pages|templates/)) {
-        const { createPage, deletePage } = actions;
-        const getTitle = (object, path) => {
-            const array = path.split('/').filter(val => val);
-            if (path === '/') {
-                array.push('home');
-            }
-            let value = {};
-            value = array.reduce(
-                (obj, key) => (obj && obj[key] !== 'undefined' ? obj[key] : undefined),
-                object,
-            );
-            return value ? value.title : 'Untitled';
-        };
-        return new Promise(resolve => {
-            deletePage(page);
-            Object.keys(i18n).map(key => {
-                return createPage({
-                    ...page,
-                    path: i18n[key].path + page.path,
-                    context: {
-                        lang: i18n[key],
-                        title: getTitle(i18n[key], page.path),
-                    },
-                });
-            });
-            resolve();
-        });
-    }
-    return null;
-};
+    // Create one page for each locale
+    await Promise.all(
+        currentLanguages.map(async lang => {
+            const originalPath = page.path
+            const localizedPath = `${lang.path}${page.path}`
 
-exports.createPages = ({ actions, graphql }) => {
-    const { createPage } = actions;
-    return graphql(`
+            await createPage({
+                ...page,
+                path: localizedPath,
+                context: {
+                    ...page.context,
+                    originalPath,
+                    lang: lang.shorthand,
+                },
+            })
+        }),
+    )
+}
+
+exports.createPages = async ({ actions, graphql, reporter }) => {
+    const { createPage } = actions
+    const projectTemplate = require.resolve(`./src/templates/project.js`)
+    const result = await graphql(`
         {
-            allMarkdownRemark(filter: { fileAbsolutePath: { regex: "/projects/" } }) {
+            allMarkdownRemark(sort: { order: DESC, fields: [frontmatter___date] }, limit: 1000) {
                 edges {
                     node {
-                        fileAbsolutePath
-                        html
                         frontmatter {
-                            title
-                            template
+                            slug
                             key
                         }
                     }
                 }
             }
         }
-    `).then(result => {
-        if (result.errors) {
-            return Promise.reject(result.errors);
-        }
-        result.data.allMarkdownRemark.edges.forEach(({ node }) => {
-            createPage({
-                path: `${i18n[node.frontmatter.key].path}/${slugify(node.frontmatter.title)}`,
-                component: path.resolve(`src/templates/${String(node.frontmatter.template)}.js`),
-                context: {
-                    filePath: node.fileAbsolutePath,
-                    lang: i18n[node.frontmatter.key],
-                    title: node.frontmatter.title,
-                    html: node.html,
-                },
-            });
-            createPage({
-                path: `${i18n[node.frontmatter.key].path}/${slugify(node.frontmatter.template)}`,
-                component: path.resolve(
-                    `src/templates/${String(node.frontmatter.template)}Index.js`,
-                ),
-                context: {
-                    key: node.frontmatter.key,
-                    lang: i18n[node.frontmatter.key],
-                    filePath: path.resolve(
-                        `src/templates/${String(node.frontmatter.template)}Index.js`,
-                    ),
-                    title: node.frontmatter.template,
-                    html: node.html,
-                },
-            });
-        });
-    });
-};
+    `)
+    // Handle errors
+    if (result.errors) {
+        reporter.panicOnBuild(`Error while running GraphQL query.`)
+        return
+    }
+    result.data.allMarkdownRemark.edges.forEach(({ node }) => {
+        const currentLanguage = currentLanguages.find(
+            language => node.frontmatter.key === language.shorthand,
+        )
+        createPage({
+            path: `${currentLanguage.path}${node.frontmatter.slug}`,
+            component: projectTemplate,
+            context: {
+                // additional data can be passed via context
+                originalPath: node.frontmatter.slug,
+                lang: currentLanguage.shorthand,
+            },
+        })
+    })
+}
