@@ -13,4 +13,4 @@ module.exports = {
     },
     testURL: `http://localhost`,
     setupFiles: [`<rootDir>/jest/loadershim.js`],
-};
+}
