@@ -1,5 +1,6 @@
-module.exports = {
-    title: 'Contact Us',
-    header: 'Contact Us',
+const contact = {
+    title: 'Contact me',
     submit: 'Submit',
-};
+}
+
+export default contact
