@@ -1,3 +1,5 @@
-module.exports = {
+const projects = {
     title: 'Projets',
-};
+}
+
+export default projects
